requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==5.0.1
Django==4.2.5
django-allauth==0.57.0
django-jazzmin==2.6.0
django-parler==2.3
django-parler-rest==2.2
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
oauthlib==3.2.2
packaging==23.1
psycopg2==2.9.8
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.5
